audio: update the onboard DAC at 1M samples/s
1kHz is 1000 samples at this sample rate.

This pretty much gets rid of all the harmonics -- they're down at -59dB.
But it seems to be a bit slow-running now; I'm measuring it at 988Hz on
the scope.

The dynamic range is dropped to ±0x700 instead of ±0x800, because the
result kept getting its tops and bottoms clipped.  This still gives me
about 2.8Vpp.