audio: fix oddity when changing frequencies
The datasheet says "It is forbidden to write these registers when the EN
bit is read as 1."

Again, I didn't expect it to read as a 1 after we just cleared it, but
this fixes some weirdness when the commands "f1000\n" followed by
"f10001\n" are sent -- it was updating the buffer correctly, but still
looping over 1000Hz-worth of samples.

It makes sense if EN remains readable as a 1 while it's still in the
process of disabling the DMA channel for us.