audio: use the USB interrupt to drive main loop
This brings the current draw down from about 68mA to 37mA while it's
running.

Getting the order of "while interrupts are disabled { check and poll()
the USB device; enable the interrupt; WFI }" was tricky, but doing it
this way lets me leave ownership in main()'s stack -- the sample buffer
in SignalGenerator is too large for me to allocate statically (which
would be necessary to use it directly from the interrupt handler) AND
temporarily on the stack while it's being initialized.