add a USB line-based Future and a tiny "executor"
This does not even compile -- I still haven't figured out how to make my
lifetime bounds line up for this.  Ideally, I want the buffer to be
borrowed for as long as the `.await` takes, but it's somehow getting
inferred to "longer than the usb_command lives" instead.

And the usb_command borrow seems to extend beyond the `.await` too!

The end result in `main()` is marginally better (if you ignore the
complexity of the Future and the executor) than it was before, and this
should let me remove the SignalGenerator from UsbCommand entirely.